aceexam_backend/requirements.txt
django>=4.2,<5.0
pymongo>=4.6
graphene-django>=3.1
django-cors-headers>=4.3
PyJWT>=2.8
